+++ exporter/collector/integrationtest/logs_test.go
@@ -19,6 +19,7 @@ import (
 	"testing"
 	"time"
 
+	"github.com/GoogleCloudPlatform/opentelemetry-operations-go/exporter/collector/internal/integrationtest/testcases"
 	"github.com/stretchr/testify/require"
 )
 
@@ -27,7 +28,7 @@ func TestLogs(t *testing.T) {
 	ctx := context.Background()
 	timestamp := time.Now()
 
-	for _, test := range LogsTestCases {
+	for _, test := range testcases.LogsTestCases {
 		test := test
 
 		t.Run(test.Name, func(t *testing.T) {
